+++ src/net/sourceforge/plantuml/jungle/PSystemTree.java
@@ -76,8 +76,8 @@ public class PSystemTree extends AbstractPSystem {
 			margin1 = 5;
 			margin2 = 5;
 		}
-		final ImageBuilder builder = ImageBuilder.buildB(new ColorMapperIdentity(), false, ClockwiseTopRightBottomLeft.margin1margin2((double) margin1, (double) margin2),
-		null, null, null, 1.0, HColorUtils.WHITE);
+		final ImageBuilder builder = ImageBuilder.buildB(new ColorMapperIdentity(), false,
+				ClockwiseTopRightBottomLeft.margin1margin2(margin1, margin2), null, null, null, 1.0, HColorUtils.WHITE);
 		if (rendering == Rendering.NEEDLE) {
 			final UDrawable tmp = Needle.getNeedle(root, 200, 0, 60);
 			final LimitFinder limitFinder = new LimitFinder(fileFormat.getDefaultStringBounder(), true);
