+++ src/net/sourceforge/plantuml/timingdiagram/TimingDiagram.java
@@ -111,8 +111,9 @@ public class TimingDiagram extends UmlDiagram implements Clocks {
 			margin1 = 10;
 			margin2 = 10;
 		}
-		final ImageBuilder imageBuilder = ImageBuilder.buildD(getSkinParam(), ClockwiseTopRightBottomLeft.margin1margin2((double) margin1, (double) margin2), getAnimation(), fileFormatOption.isWithMetadata() ? getMetadata() : null,
-		getWarningOrError(), dpiFactor);
+		final ImageBuilder imageBuilder = ImageBuilder.buildD(getSkinParam(),
+				ClockwiseTopRightBottomLeft.margin1margin2(margin1, margin2), getAnimation(),
+				fileFormatOption.isWithMetadata() ? getMetadata() : null, getWarningOrError(), dpiFactor);
 
 		TextBlock result = getTextBlock();
 		final ISkinParam skinParam = getSkinParam();
