+++ src/net/sourceforge/plantuml/wire/WireDiagram.java
@@ -83,7 +83,7 @@ public class WireDiagram extends UmlDiagram {
 
 		final double dpiFactor = scale == null ? getScaleCoef(fileFormatOption) : scale.getScale(100, 100);
 		final ISkinParam skinParam = getSkinParam();
-		
+
 		final int margin1;
 		final int margin2;
 		if (SkinParam.USE_STYLES()) {
@@ -93,8 +93,10 @@ public class WireDiagram extends UmlDiagram {
 			margin1 = 10;
 			margin2 = 10;
 		}
-		final ImageBuilder imageBuilder = ImageBuilder.buildB(skinParam.getColorMapper(), skinParam.handwritten(), ClockwiseTopRightBottomLeft.margin1margin2((double) margin1, (double) margin2),
-		null, fileFormatOption.isWithMetadata() ? getMetadata() : null, "", dpiFactor, skinParam.getBackgroundColor(false));
+		final ImageBuilder imageBuilder = ImageBuilder.buildB(skinParam.getColorMapper(), skinParam.handwritten(),
+				ClockwiseTopRightBottomLeft.margin1margin2(margin1, margin2), null,
+				fileFormatOption.isWithMetadata() ? getMetadata() : null, "", dpiFactor,
+				skinParam.getBackgroundColor(false));
 		TextBlock result = getTextBlock();
 
 		result = new AnnotatedWorker(this, skinParam, fileFormatOption.getDefaultStringBounder()).addAdd(result);
